requirements.txt
pandas>=2.2
openpyxl>=3.1
pyyaml>=6.0
pydantic>=2.8
typer>=0.12
matplotlib>=3.8
